feat: Implement graph models with external feats
This commit adds the following implementations to the models subpackage:
- GraphAttentionNetworkWithExternal
- GraphConvolutionalNetworkWithExternal
- GraphSAGENetworkWithExternal
- GraphTransformerNetworkWithExternal

In these models the corresponding graph neural network is employed to produce
global level representations from smiles. Then these are concatenated with the
external feature vectors and the concatenated vector is passed through a fully
connected network to produce the final output.