feat(JAQPOT-127):  jaqpotpy-datasets-refactoring (#28)
* fix: Deleted datasets

All test datasets should be concentrated at a single point and not in several subfolders within jaqpotpy

* fix: super minor changes

* fix: Old abstract dataset classes

This file is going to be deleted once the new classes are up and running

* feat: New abtract classes

* feat: old molecular datasets

This file is going to be deleted once the new dataset classes are ready

* feat: deleted Molecular dataset

This class was completely redundant

* feat: Added error handling in csv reading

* feat: added dataframe setter, getter, deleter

* fix: Deleted redundant comments

* feat: Deleted SmilesDataset class

* feat: error handling for y_cols

* feat: x_cols error handling

* feat: smiles_cols error handling

* feat: Handling more that one smiles column

* feat: Finalized Jaqpotpy dataset class

All commints after this commit will fix bugs

* fix: Smiles parsing

There was an error when the user was providing a list of a single string. Now it is fixed

* feat: Making sure that the user provided y_cols correctly

* fix: delete redundant code

Instead of taking different scenarios, just drop smiles into a list and work with this throughout the code

* fix: deleted unessecary code

* feat: added extra features to test datasets

* feat: Implemented __get__X and __get__Y methods

* chore: removed module loading in dataset inits

* chore: Minor changes in test datasets

* feat: Make sure there is no overlap between x_cols, y_cols and smiles_cols

* feat: Check that all user-specified cols exist

* test: early-dataset-testing

This consists of the first round of tests for dataset construction. Further tests need to be written, this is a start

* refactor: more robust error handling

* test: Added further dataset tests

* test: df and path test

Added a test that checks if a TypeError is raised in the case of the user providing both a path and a csv

* chore: Introduce JapoqtpyDataset

Change all references of SmilesDataset and MolecularTabularDataset to JaqpotpyDataset

* fix: module loading issue

* chore: changed MolecularDataset with JaqpotpyDataset

* chore: minor dependency fix

* chore: minor dependency fix

* chore: minor dependency fix

* chore: Restored dataset.__init__.py

* chore: removed wildcard import

Removed the wildcard import from datasets.__inits__

* chore: renamed doa_m to doa_fitted

doa_m was the assigned to doa.fit() so this was renamed accordingly

* fix: Compatibility between MolecularSKLearn and JaqpotpyDataset

The changes on JaqpotpyDataset had created several compatibilty issues with the classes MolecularModel() and MolecularSKLearn which are now fixed. Next thing to fix is the .predict() method and also to checkout what is going on with the onnx framework and the predictions

* chore: rename Preprocesses class to Preprocess

* chore: Rename Preprocesses to Preprocess in all files

* feat: Add a .copy() method in JaqpotpyDataset

* chore: Return pandas series instead of numpy with __get__X__ and y

* fix: Now the .copy() method works

* chore: comment with tests to be added

* refactor: Preprocessing made an object from class, attribute

Previously, once a preprocessor was defined, preprocessing of X and y were considered class attributes instead of object attributes, so no changes could be made. Now preprocessing of X and y are object attributes and as such are initialised and can be reset with each new object initilisation

* Refactor: Major model refactor

models are now compatible with the new JaqpotpyDataset

* chore: removed references to Molecular SKLearn

* fix: Fixed error in len(dataset)